missing_persons_backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow>=10.2.0
opencv-python==4.8.1.78
numpy>=1.24.3
scipy>=1.11.4
deepface>=0.0.75
ultralytics>=8.0.228
python-decouple==3.8
celery==5.3.4
redis==5.0.1
